backend/data-collection-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
beautifulsoup4>=4.12.0
requests>=2.31.0
boto3>=1.26.0
python-dotenv>=1.0.0